GH-113568: Stop raising deprecation warnings from pathlib ABCs
